hf2dataset.cpp: Cast nUTMZone to an int as std::abs converted a 16 bit int GInt16 to a float (https://trac.osgeo.org/gdal/ticket/6700).
git-svn-id: https://svn.osgeo.org/gdal/trunk@35960 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/hf2/hf2dataset.cpp
@@ -613,7 +613,7 @@ GDALDataset *HF2Dataset::Open( GDALOpenInfo * poOpenInfo )
         if (bHasUTMZone && std::abs(nUTMZone) >= 1 && std::abs(nUTMZone) <= 60)
         {
             bHasSRS = true;
-            oSRS.SetUTM(std::abs(nUTMZone), nUTMZone > 0);
+            oSRS.SetUTM(std::abs(static_cast<int>(nUTMZone)), nUTMZone > 0);
         }
         if (bHasSRS)
             oSRS.exportToWkt(&poDS->pszWKT);
